+++ crates/astria-sequencer/src/app/tests_app/mempool.rs
@@ -289,11 +289,9 @@ async fn maintenance_recosting_promotes() {
         "two txs in mempool; one included in proposal is not yet removed"
     );
 
-    // set dummy hash
-    app.executed_proposal_hash = Hash::try_from([97u8; 32].to_vec()).unwrap();
-
+    let hash = Hash::Sha256([97u8; 32]);
     let process_proposal = abci::request::ProcessProposal {
-        hash: app.executed_proposal_hash,
+        hash,
         height: Height::default(),
         time: Time::now(),
         next_validators_hash: Hash::default(),
@@ -314,7 +312,7 @@ async fn maintenance_recosting_promotes() {
 
     // finalize with finalize block
     let finalize_block = abci::request::FinalizeBlock {
-        hash: app.executed_proposal_hash,
+        hash,
         height: 1u32.into(),
         time: Time::now(),
         next_validators_hash: Hash::default(),
@@ -461,9 +459,9 @@ async fn maintenance_funds_added_promotes() {
         "only one transactions should've been valid (besides 2 generated txs)"
     );
 
-    app.executed_proposal_hash = Hash::try_from([97u8; 32].to_vec()).unwrap();
+    let hash = Hash::Sha256([97u8; 32]);
     let process_proposal = abci::request::ProcessProposal {
-        hash: app.executed_proposal_hash,
+        hash,
         height: Height::default(),
         time: Time::now(),
         next_validators_hash: Hash::default(),
@@ -482,12 +480,9 @@ async fn maintenance_funds_added_promotes() {
     yet removed"
     );
 
-    // set dummy hash
-    app.executed_proposal_hash = Hash::try_from([97u8; 32].to_vec()).unwrap();
-
     // finalize with finalize block
     let finalize_block = abci::request::FinalizeBlock {
-        hash: app.executed_proposal_hash,
+        hash,
         height: 1u32.into(),
         time: Time::now(),
         next_validators_hash: Hash::default(),